split e2e tests in python and ts
